+++ src/base/generator/base_generator.py
@@ -1,8 +1,8 @@
-import src_2.base.configuration
-import src_2.base.generator.files
+import src.base.configuration
+import src.base.generator.files
 
 
-class BaseGenerator(src_2.base.generator.files.FilesGenerator):
+class BaseGenerator(src.base.generator.files.FilesGenerator):
     def generate(self):
         return {
             "text_files": self.generate_non_media_files(),
