+++ src/base/generator/images.py
