+++ src/base/logs.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
 """Handle logs information
 
 Relevant Global Variables:
@@ -12,25 +11,52 @@ Classes:
     Logs
     SpecialMessages
 """
-
 import sys
-
-from src.helpers.strings import ColorProcessor
-
 import textwrap
 
-from src.info import Info
+import src.info
+import src.support
 
 
-_INFO = Info.get_info()
+class ColorProcessor:
+    """Class to handle the color processing
+
+    Init:
+        string str = ''
+
+    Methods:
+        default_color
+        error_color
+        presentation_color
+    """
+
+    def __init__(self, string: str = ""):
+        self.string = string
+
+    def default_color(self) -> str:
+        """Return a string with the default color"""
+        return self.string
+
+    def error_color(self) -> str:
+        """Return a string with error color"""
+        return (src.support.ERROR_COLOR + self.string +
+                src.support.DEFAULT_COLOR)
+
+    def presentation_color(self) -> str:
+        """Return a string with presentation color"""
+        return (src.support.PRESENTATION_COLOR + self.string +
+                src.support.DEFAULT_COLOR)
+
+
+_INFO = src.info.get_info()
 
 # presentation is in /docs/presentation.png
 PRESENTATION_MESSAGE = textwrap.dedent("""\
        ____  _   _  ____   _   _  _____     _     ____     ____ __   __
       / ___|| | | |/ ___| | | | || ____|   / \\   |  _ \\   |  _ \\\\ \\ / /
-     | |    | | | |\\___ \\ | |_| ||  _|    / _ \\  | | | |  | |_) |\\ V / 
-     | |___ | |_| | ___) ||  _  || |___  / ___ \\ | |_| |_ |  __/  | |  
-      \\____| \\___/ |____/ |_| |_||_____|/_/   \\_\\|____/(_)|_|     |_|  
+     | |    | | | |\\___ \\ | |_| ||  _|    / _ \\  | | | |  | |_) |\\ V /
+     | |___ | |_| | ___) ||  _  || |___  / ___ \\ | |_| |_ |  __/  | |
+      \\____| \\___/ |____/ |_| |_||_____|/_/   \\_\\|____/(_)|_|     |_|
                                  __       _
                                  _/     /
                                 /    __/
@@ -49,14 +75,14 @@ PRESENTATION_MESSAGE = textwrap.dedent("""\
     For help run: {} -h
     """)  # This line is blank
 PRESENTATION_MESSAGE = PRESENTATION_MESSAGE.format(
-    _INFO['package_version'],
-    _INFO['author'],
-    _INFO['email'],
-    _INFO['author_page'],
-    _INFO['license'],
-    _INFO['source'],
-    _INFO['documentation'],
-    _INFO['package_name'],
+    _INFO["package_version"],
+    _INFO["author"],
+    _INFO["email"],
+    _INFO["author_page"],
+    _INFO["license"],
+    _INFO["source"],
+    _INFO["documentation"],
+    _INFO["package_name"],
 )
 
 
@@ -73,55 +99,43 @@ class MessagesHandler:
     # default section
 
     @staticmethod
-    def default_stdout(message: str = ''):
-        class_instance = ColorProcessor(message + '\n')
+    def default_stdout(message: str = ""):
+        class_instance = ColorProcessor(message + "\n")
         sys.stdout.write(class_instance.default_color())
 
     # error section
 
     @staticmethod
-    def error_exception(message: str = ''):
+    def error_exception(message: str = ""):
         raise Exception(message)
 
     @staticmethod
-    def error_stdout(message: str = ''):
+    def error_stdout(message: str = ""):
         class_instance = ColorProcessor(message)
         sys.exit(class_instance.error_color())
 
     # presentation section
 
     @staticmethod
-    def presentation_stdout(message: str = ''):
-        class_instance = ColorProcessor(message + '\n')
+    def presentation_stdout(message: str = ""):
+        class_instance = ColorProcessor(message + "\n")
         sys.stdout.write(class_instance.presentation_color())
 
 
 class Logs(MessagesHandler):
     """Different output handler for different situations
-    
+
     Methods:
         default_log
         error_log
         presentation_log
     """
 
-    def default_log(self, message: str = ''):
+    def default_log(self, message: str = ""):
         self.default_stdout(message)
 
-    def error_log(self, message: str = ''):
+    def error_log(self, message: str = ""):
         self.error_exception(message)
 
-    def presentation_log(self, message: str = ''):
+    def presentation_log(self, message: str = ""):
         self.presentation_stdout(message)
-
-
-class SpecialMessages:
-    """Special messages class
-
-    Methods:
-        @staticmethod presentation_message
-    """
-
-    @staticmethod
-    def presentation_message():
-        return PRESENTATION_MESSAGE
