+++ src/console/arguments.py
@@ -1,30 +1,21 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-"""Module used to handle the argparse library
-
-Classes:
-    Argparse
-"""
 import argparse
 import textwrap
-from argparse import Namespace
-from typing import Union
+import typing
 
-from src.helpers.assets import Images
-from src.helpers.strings import Transformator
-from src.helpers.validators import FilesValidator
-from src.info import Info
-from src.services.logs import Logs
+import src.base.logs
+import src.helpers
+import src.info
 
 
-class Argparse(Info, Logs):
+class Argparse(src.base.logs.Logs):
     """Class used to handle argparse
 
     Methods:
         parse_args
     """
 
-    def parse_args(self, args: Union[list, None] = None) -> Namespace:
+    def parse_args(self, args: typing.Union[list, None] = None
+                   ) -> argparse.Namespace:
         """Argparse implementation
 
         This function validates the values of arguments and, if everything is
@@ -33,20 +24,18 @@ class Argparse(Info, Logs):
         if not isinstance(args, list):
             args = []
 
-        info = self.get_info()
+        info = src.info.get_info()
         name = info["package_name"]
         parser = argparse.ArgumentParser(
             prog=info["package_name"],
             formatter_class=argparse.RawDescriptionHelpFormatter,
             usage=f"{info['package_name']} -config FILEPATH",
-            epilog=textwrap.dedent(
-                f"""\
+            epilog=textwrap.dedent(f"""\
                 Examples:
                 1) Generate default config file with images:
                     {name} -default settings.json --images
                 2) Run that config:
-                    {name} -config settings.json"""
-            ),
+                    {name} -config settings.json"""),
         )
 
         # ARGUMENTS
@@ -60,10 +49,8 @@ class Argparse(Info, Logs):
             "-config",
             metavar="FILEPATH",
             dest="config",
-            help=(
-                "Path to a config file in JSON format. "
-                "Read a config file and create the main files based on it."
-            ),
+            help=("Path to a config file in JSON format. "
+                  "Read a config file and create the main files based on it."),
         )
         # -default
         main_arguments.add_argument(
@@ -71,23 +58,19 @@ class Argparse(Info, Logs):
             metavar="FILENAME",
             dest="default",
             help=(
-                "Path to output a default config file. Can use with --images"
-            ),
+                "Path to output a default config file. Can use with --images"),
         )
         # GROUP: options
         # -images
-        image_list = Images.images_list()
-        class_instance = Transformator(string_list=image_list)
-        joined_words = class_instance.string_list_union()
+        images_list = src.helpers.images_list()
+        joined_words = src.helpers.string_list_union(images_list)
         complementary_arguments.add_argument(
             "--images",
             dest="images",
             action="store_true",
-            help=(
-                f"Use with -default. "
-                f"Generate default images that can be used by the settings. "
-                f"This include: {joined_words}"
-            ),
+            help=(f"Use with -default. "
+                  f"Generate default images that can be used by the settings. "
+                  f"This include: {joined_words}"),
         )
 
         # Set defaults
@@ -101,19 +84,18 @@ class Argparse(Info, Logs):
         # pylint no-member error in child function
         parsed_args = parser.parse_args(args)
         if not (parsed_args.config or parsed_args.default):
-            self.error_log("Miss Required arguments. Use -config or -default")
+            self.error_log(
+                "Miss Required arguments. Use -config or -default. Use -h for help"
+            )
         if parsed_args.config and parsed_args.default:
             self.error_log(
-                "Can't use -config and -default arguments together."
-            )
+                "Can't use -config and -default arguments together.")
         if parsed_args.images and not parsed_args.default:
             self.error_log("Can't use --images without -default.")
         if parsed_args.config:
-            validator = FilesValidator(
-                file_path=parsed_args.config, key="-config"
-            )
-            validation = validator.path_is_not_directory()
-            if validation:
-                self.error_log(validation)
+            error = src.helpers.path_is_not_directory(
+                parsed_args.config, "-config")
+            if error:
+                self.error_log(error)
 
         return parsed_args
