--- src_2/cushead.py
@@ -1,15 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-import src_2.base.configuration
-import src_2.base.generator.base_generator
-import src_2.support
-
-
-def generate_files(config):
-    src_2.support.Support().check_for_execution()
-    src_2.base.generator.base_generator.BaseGenerator(config).generate_files()
-
-
-def generate_default_config(output_path, add_assets=False):
-    src_2.support.Support().check_for_execution()
-    src_2.base.configuration.generate_default_config(output_path, add_assets)
